test sndfft

+++ wegi/egi_math.h
@@ -67,7 +67,7 @@ inline EGI_FCOMPLEX 	mat_CompMult(EGI_FCOMPLEX a, EGI_FCOMPLEX b);
 inline EGI_FCOMPLEX 	mat_CompDiv(EGI_FCOMPLEX a, EGI_FCOMPLEX b);
 float 			mat_floatCompAmp( EGI_FCOMPLEX a );
 unsigned int 		mat_uintCompAmp( EGI_FCOMPLEX a );
-unsigned int 		mat_uintCompSAmp( EGI_FCOMPLEX a );
+uint64_t 		mat_uintCompSAmp( EGI_FCOMPLEX a );
 EGI_FCOMPLEX 		*mat_CompFFTAng(uint16_t np);
 int 			mat_egiFFFT( uint16_t np, const EGI_FCOMPLEX *wang,
                                      const float *x, const int *nx, EGI_FCOMPLEX *ffx);
